(feat): Created the modules and reorganized the project
The project has been reorganized with all the modules and the old models have been moved
