+++ FitWin/FitWin/settings.py
@@ -39,7 +39,11 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
 
-    'anuncios',
+    'authentication',
+    'announcements',
+    'payment',
+    'searching',
+    'users',
 
     'landingPage',
 ]
@@ -117,7 +121,7 @@ AUTH_PASSWORD_VALIDATORS = [
 
 LANGUAGE_CODE = 'en-us'
 
-TIME_ZONE = 'UTC'
+TIME_ZONE = 'CET'
 
 USE_I18N = True
