+++ FitWin/announcements/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
 from .models import *
 
-admin.site.register(Trainer)
-admin.site.register(Client)
 admin.site.register(Announcement)
 admin.site.register(Category)
