+++ FitWin/announcements/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AnunciosConfig(AppConfig):
+class AnnouncementsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'anuncios'
+    name = 'announcements'
