+++ FitWin/announcements/models.py
@@ -1,32 +1,11 @@
 from django.db import models
-from django.contrib.auth.models import User
-# Create your models here.
-
-
-class Trainer(models.Model):
-    
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    picture=models.ImageField(blank=True, null=True)
-    date_created = models.DateField(auto_now_add=True)
-    birthday=models.DateField(null=True, blank=True)
-    bio = models.TextField(max_length=150, null=True, blank=True)
-
-
-class Client(models.Model):
-    
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    picture=models.ImageField(blank=True, null=True)
-    date_created = models.DateField(auto_now_add=True)
-    birthday=models.DateField(null=True, blank=True)
-    bio = models.TextField(max_length=150, null=True, blank=True)
+from users.models import Trainer, Client
 
 class Category(models.Model):
     name=models.CharField(max_length=250, verbose_name='Categoria')
 
     def __str__(self):
         return self.name
-
-
 
 class Announcement(models.Model):
     title = models.CharField(max_length=250, verbose_name='Titulo')
