+++ FitWin/authentication/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('login', views.login, name='login'),
+
+]
